reworks/wizard
so I noticed some issues with some of the models textures and so I reworked those and tried as best as i could to get them fixed. I put in the animaitons for the wizard i dont have al them connected cuz i didnt know how people wanted it so i didnt touch it besides an idle.